mindmap/requirements.txt
flask
python-dotenv
transformers
torch
sentencepiece
huggingface-hub
gunicorn